requirements.txt
python-telegram-bot
librosa
numpy